pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "expand_rgb"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "click>=8.1.7",
    "numpy>=1.24.4",
    "pillow>=10.4.0",
    "tqdm>=4.66.5",
]
[tool.maturin]
features = ["pyo3/extension-module"]
